Replace stream-processor-fold framework with a plain stream-view reducer
The vendored legacy processor framework (821 lines) existed only to run one
contract with empty events/consumes/emits + consumesAllEvents. The reduce
body moves verbatim into a plain (state, event) => state fold; a minimal
StreamEvent type replaces the zod-heavy vendored model.
